requirements.txt
# Predictive Maintenance System - Complete Requirements
# Install all dependencies: pip install -r requirements.txt
# Python 3.9+ recommended

# =============================================================================
# CORE DEPENDENCIES
# =============================================================================

# Scientific Computing
numpy>=1.24.3
pandas>=2.1.4
scipy>=1.11.3

# Configuration
pyyaml>=6.0.1

# =============================================================================
# DATA INFRASTRUCTURE
# =============================================================================

# Kafka Client (Data Generator & Stream Processor)
kafka-python>=2.0.2

# Database Drivers
psycopg2-binary>=2.9.9  # PostgreSQL/TimescaleDB
redis>=5.0.1  # Redis cache

# =============================================================================
# MACHINE LEARNING
# =============================================================================

# Deep Learning
tensorflow>=2.13.0
keras>=2.15.0

# Classical ML
scikit-learn>=1.3.2
xgboost>=2.0.3

# Experiment Tracking & Model Management
mlflow>=2.9.2

# Utilities
joblib>=1.3.2
tqdm>=4.66.1

# =============================================================================
# API & WEB SERVICES
# =============================================================================

# FastAPI (Inference Service)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# HTTP Client
requests>=2.31.0

# =============================================================================
# DASHBOARDS & VISUALIZATION
# =============================================================================

# Streamlit Dashboard
streamlit>=1.31.0
streamlit-autorefresh>=1.0.1

# Plotting Libraries
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0

# =============================================================================
# MONITORING & OBSERVABILITY
# =============================================================================

# Prometheus Metrics
prometheus-client>=0.19.0

# Structured Logging
python-json-logger>=2.0.7

# =============================================================================
# UTILITIES
# =============================================================================

python-dateutil>=2.8.2

# =============================================================================
# DEVELOPMENT & TESTING (Optional)
# =============================================================================

# Testing Framework
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
httpx>=0.25.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.27.0

# =============================================================================
# OPTIONAL: Email Notifications (Alert Engine)
# =============================================================================
# Uncomment if using email alerts:
# secure-smtplib>=0.1.1

# =============================================================================
# OPTIONAL: Additional Database Connectors
# =============================================================================
# Uncomment as needed:
# pymongo>=4.6.1  # MongoDB
# sqlalchemy>=2.0.25  # ORM
# influxdb-client>=1.39.0  # InfluxDB

# =============================================================================
# INSTALLATION NOTES
# =============================================================================
#
# Full Installation (All Components):
#   pip install -r requirements.txt
#
# Minimal Installation (Core Only):
#   pip install numpy pandas pyyaml scikit-learn
#
# Development Installation (with testing tools):
#   pip install -r requirements.txt
#   # All dev dependencies included
#
# Module-Specific Installation:
#   If you only need specific modules, see individual requirements.txt in:
#   - data_generator/requirements.txt
#   - stream_processor/requirements.txt
#   - feature_store/requirements.txt
#   - ml_pipeline/train/requirements.txt
#   - ml_pipeline/evaluate/requirements.txt
#   - ml_pipeline/retrain/requirements.txt
#   - inference_service/requirements.txt
#   - alerting/requirements.txt
#   - dashboard/requirements.txt
#
# =============================================================================
# SYSTEM REQUIREMENTS
# =============================================================================
#
# Python: 3.9+
# RAM: 8GB minimum, 16GB recommended
# Storage: 10GB+ for models and data
# OS: Linux (Ubuntu 20.04+), macOS 11+, Windows 10+ with WSL2
#
# Docker & Docker Compose (for infrastructure):
#   - Docker Engine 20.10+
#   - Docker Compose 2.0+
#
